Add product unit (kg/g/l/ml) and store measurements as decimals
Products now require a unit and store stock_quantity as Numeric(10,3)
instead of an integer, so fractional amounts (e.g. 2.5 kg) are
representable. Order item quantity follows the same change so
customers can order fractional amounts of measured products. Existing
products were backfilled to unit='kg' during migration; existing
integer stock/quantity values converted losslessly to decimal.

+++ app/models/order.py
@@ -3,7 +3,7 @@ import uuid
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import DateTime, Enum, ForeignKey, Integer, Numeric, String, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Numeric, String, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -50,7 +50,7 @@ class OrderItem(Base):
     product_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("products.id", ondelete="RESTRICT"), nullable=False
     )
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    quantity: Mapped[Decimal] = mapped_column(Numeric(10, 3), nullable=False)
     price_at_purchase: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
 
     order: Mapped["Order"] = relationship(back_populates="items")
